Split the cbenchio files in two files
