+++ tests/synth/cbenchio/cbenchio.py
@@ -30,24 +30,9 @@ import reframe.core.meta as meta
 
 Used for I/O benchmarking.
 
-## Adding new tests
-
-Add a new test by inherting from cbenchio_bandwidthBase and using the Parameterize metaclass. The class should have a config attribute pointing to a yaml file. The yaml file should contain the parameters for the test. For example:
-
-```python
-class cbenchio_bandwidth_write(cbenchio_bandwidthBase,metaclass=Parameterize):
-    operation = "write"
-    config= "posix.yaml"
-```
-
-Read tests can be generated from write tests using the make_read_test function. For example we can generate a read test from the write test above using:
-
-```python
-cbenchio_bandwidth_read = make_read_test(cbenchio_bandwidth_write)
-```
+Base and meta classes for cbenchio benchmarks
 
 """
-
 
 def get_config(filename: str) -> dict:
 
@@ -240,31 +225,12 @@ class cbenchio_read(cbenchio):
             print(f"Warning: Failed to clean up directory {self.path}: {e}")
 
     
-
 def make_read_test(cls):
     
     # check that the class contains write
     if cls.__name__.find("write") == -1:
         raise ValueError("The class passed to make_read_test must contain 'write' in its name")
     
-    return rfm.simple_test(rfm.core.meta.make_test(cls.__name__.replace("write", "read"), (cbenchio_read,), {"operation": "read","write_test": fixture(cls, scope='environment'),}, module=__name__))
-    
-
-class cbenchio_posix_sequential_write(cbenchio_write,metaclass=Parameterize):
-    """ Measure the bandwidth of the filesystem for file per process patterns for large sequential I/O. """
-    operation = "write"
-    config= "posix-large-sequential.yaml"
-
-cbenchio_posix_sequential_read = make_read_test(cbenchio_posix_sequential_write)
-
-
-
-class cbenchio_posix_random_write(cbenchio_write,metaclass=Parameterize):
-    """ Measure the bandwidth of the filesystem for file per process patterns random 4KiB I/O. """
-    operation = "write"
-    config= "posix-large-random.yaml"
-    random_strided = True
-    max_random_stride = 4096 # KiB
-    file_size_per_process = 4 # MiB
-
-cbenchio_posix_random_read = make_read_test(cbenchio_posix_random_write)
+    fixture = rfm.core.builtins.fixture(cls, scope='environment')
+    module=fixture.cls.__module__
+    return rfm.simple_test(rfm.core.meta.make_test(cls.__name__.replace("write", "read"), (cbenchio_read,), {"operation": "read","write_test": fixture,}, module=module) )
